refactor(moddy): split commands into modules
