+++ moddy/src/__init__.py
@@ -1,0 +1,13 @@
+AUTO_YES = False
+MODDY_VERSION = "DEVELOPMENT"
+VERSION_REGISTRY_URL = (
+    "https://raw.githubusercontent.com/iamkaf/modresources/main/moddy/versions.json"
+)
+RAW_BASE_URL = "https://raw.githubusercontent.com/iamkaf/modresources/main"
+
+__all__ = [
+    "AUTO_YES",
+    "MODDY_VERSION",
+    "VERSION_REGISTRY_URL",
+    "RAW_BASE_URL",
+]
