+++ moddy/src/commands/__init__.py
@@ -1,0 +1,17 @@
+from .add_service import cmd_add_service
+from .open_libs import cmd_open_libs
+from .set_minecraft_version import cmd_set_minecraft_version
+from .setup_template import cmd_setup
+from .update import cmd_update
+from .meta import cmd_help, cmd_version, check_for_update
+
+__all__ = [
+    "cmd_add_service",
+    "cmd_open_libs",
+    "cmd_set_minecraft_version",
+    "cmd_setup",
+    "cmd_update",
+    "cmd_help",
+    "cmd_version",
+    "check_for_update",
+]
