+++ moddy/src/commands/update.py
@@ -1,0 +1,56 @@
+from __future__ import annotations
+
+import argparse
+import json
+import re
+import shutil
+from pathlib import Path
+
+from ..utils import fetch_url_text
+from .. import AUTO_YES, MODDY_VERSION, VERSION_REGISTRY_URL, RAW_BASE_URL
+
+
+def cmd_update(args: argparse.Namespace) -> None:
+    """Download the latest version of Moddy and replace this file."""
+    print("WARNING: this will download and execute code from the internet.")
+    try:
+        registry = json.loads(fetch_url_text(VERSION_REGISTRY_URL))
+        latest = registry[0]
+        update_url = RAW_BASE_URL + latest.get("source", "")
+    except Exception as e:
+        print(f"Failed to check for updates: {e}")
+        return
+
+    print(f"Registry: {VERSION_REGISTRY_URL}")
+    print(f"Source: {update_url}")
+    if not AUTO_YES and input("Are you sure you want to continue? [y/N] ").lower() != "y":
+        print("Aborted")
+        return
+
+    try:
+        new_code = fetch_url_text(update_url)
+    except Exception as e:
+        print(f"Failed to download update: {e}")
+        return
+
+    m = re.search(r"MODDY_VERSION\s*=\s*['\"]([^'\"]+)['\"]", new_code)
+    new_version = m.group(1) if m else "unknown"
+    if new_version == MODDY_VERSION:
+        print("Moddy is already up to date.")
+        return
+
+    script_path = Path(__file__).resolve()
+    backup = script_path.with_suffix(".bak")
+    try:
+        shutil.copy2(script_path, backup)
+        script_path.write_text(new_code, encoding="utf-8")
+    except Exception as e:
+        print(f"Update failed: {e}")
+        return
+    print(f"Updated Moddy from {MODDY_VERSION} to {new_version}")
+    print(f"A backup of the previous version was saved to {backup}")
+    notes = latest.get("notes", []) if isinstance(latest, dict) else []
+    if notes:
+        print(f"Changelog for {new_version}:")
+        for n in notes:
+            print(f" - {n}")
